test(gi): move real-model evidence-stack tests to e2e (3-tier policy)
ci-fast's check-test-policy flagged @pytest.mark.ml_models in an integration
test — real ML models belong in tests/e2e/ per the 3-tier ML/AI testing policy
(docs/architecture/TESTING_STRATEGY.md). The evidence-stack un-mock had put the
real embedding/QA/NLI tests in the integration file.

Split: the real-model tests move to tests/e2e/test_evidence_stack_e2e.py
(@e2e @ml_models, skip-when-unprovisioned); the mocked threshold/error-handling
wiring tests stay in tests/integration/gi/test_evidence_stack_integration.py
(@integration). Behaviour/assertions unchanged — only the tier placement.
